requirements.txt
Flask==0.8
Jinja2==2.6
flask-github==0.1.3
gevent==0.13.7
twitter==1.7.2
juggernaut==0.2
celery==2.5.3
Flask-Celery==2.4.2
redis==2.4.12
python-instagram==0.7.3